feat: move crypto-portfolio-metadata package to multi-layer package
