--- ur-packages/crypto-portfolio-metadata/CHANGELOG.md
@@ -1,20 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## 0.1.2-beta.0 (2024-10-16)
-
-### Bug Fixes
-
-- sync main branch pipeline trigger ([571cf94](https://github.com/ngraveio/ur-registry/commit/571cf94522a07f92ee7ee94bcd81f3cae9741b1c))
-
-## [0.1.1](https://github.com/ngraveio/ur-registry/compare/@ngraveio/bc-ur-registry-crypto-portfolio-metadata@0.1.0...@ngraveio/bc-ur-registry-crypto-portfolio-metadata@0.1.1) (2023-06-06)
-
-**Note:** Version bump only for package @ngraveio/bc-ur-registry-crypto-portfolio-metadata
-
-# 0.1.0 (2023-06-06)
-
-### Features
-
-- **metadata:** Move portfolio-metadata to its own package ([46e9952](https://github.com/ngraveio/ur-registry/commit/46e9952110cd007a3b119951b7e9fdebfc63fed9))
