--- ur-packages/crypto-portfolio-metadata/jest.config.js
@@ -1,14 +1,0 @@
-module.exports = {
-  preset: 'ts-jest',
-  testTimeout: 20000,
-  testEnvironment: 'node',
-  testPathIgnorePatterns: ['.*/node_modules/'],
-  modulePathIgnorePatterns: ['./dist'],
-  //setupFiles: ['dotenv/config'],
-  coverageReporters: ["cobertura", "clover", "json", "lcov", "text"],
-  collectCoverageFrom: [
-    '**/*.{ts,jsx}',
-    '!**/node_modules/**',
-    '!**/vendor/**',
-  ],
-};
