--- ur-packages/crypto-portfolio-metadata/package.json
@@ -1,53 +1,0 @@
-{
-  "name": "@ngraveio/bc-ur-registry-crypto-portfolio-metadata",
-  "version": "1.0.0-beta.0",
-  "description": "Provides BC UR type for crypto portfolio metadata.",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "directories": {
-    "lib": "src",
-    "test": "__tests__"
-  },
-  "files": [
-    "dist"
-  ],
-  "sideEffects": ["dist/addToRegistry.js"],
-  "scripts": {
-    "clean": "rm -rf ./dist",
-    "watch": "tsc --build --watch",
-    "build": "tsc --build",
-    "test": "jest --passWithNoTests"
-  },
-  "author": "Irfan Bilaloglu<irfan.bilaloglu@ngrave.io>",
-  "license": "MIT",
-  "homepage": "https://github.com/ngraveio/ur-registry",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ngraveio/ur-registry.git",
-    "directory": "ur-packages/crypto-portfolio-metadata"
-  },
-  "keywords": [
-    "bc-ur",
-    "Blockchain Commons",
-    "cbor",
-    "crypto",
-    "coin",
-    "portfolio"
-  ],
-  "bugs": {
-    "url": "https://github.com/ngraveio/ur-registry/issues"
-  },
-  "dependencies": {
-    "@ngraveio/bc-ur": "2.0.0-beta.6",
-    "buffer": "^6.0.3"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.1",
-    "jest": "^29.5.0",
-    "ts-jest": "^29.1.0",
-    "typescript": "^5.6.2"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
