--- ur-packages/crypto-portfolio-metadata/src/addToRegistry.ts
@@ -1,4 +1,0 @@
-import { UrRegistry } from '@ngraveio/bc-ur'
-import { PortfolioMetadata } from './PortfolioMetadata';
-
-UrRegistry.addItem(PortfolioMetadata)
