--- ur-packages/crypto-portfolio-metadata/src/index.ts
@@ -1,3 +1,0 @@
-import './addToRegistry'
-
-export * from './PortfolioMetadata'
