--- ur-packages/crypto-portfolio-metadata/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "include": ["src/*"],
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist",
-  },
-  "references": [],
-}
