+++ ur-packages/multi-layer-sync/package.json
@@ -39,10 +39,10 @@
     "url": "https://github.com/ngraveio/ur-registry/issues"
   },
   "dependencies": {
+    "buffer": "^6.0.3",
     "@ngraveio/bc-ur": "2.0.0-beta.6",
     "@ngraveio/bc-ur-registry": "1.0.0-beta.0",
     "@ngraveio/bc-ur-registry-crypto-coin-identity": "1.0.0-beta.0",
-    "@ngraveio/bc-ur-registry-crypto-portfolio-metadata": "1.0.0-beta.0",
     "@ngraveio/bc-ur-registry-hex-string": "1.0.0-beta.0"
   },
   "devDependencies": {
