+++ ur-packages/multi-layer-sync/src/addToRegistry.ts
@@ -1,8 +1,10 @@
 import { UrRegistry } from '@ngraveio/bc-ur'
+import { PortfolioMetadata } from './PortfolioMetadata'
 import { DetailedAccount } from './DetailedAccount'
 import { PortfolioCoin } from './PortfolioCoin'
 import { Portfolio } from './Portfolio'
 
+UrRegistry.addItem(PortfolioMetadata)
 UrRegistry.addItem(DetailedAccount)
 UrRegistry.addItem(PortfolioCoin)
 UrRegistry.addItem(Portfolio)
