+++ ur-packages/multi-layer-sync/src/index.ts
@@ -1,5 +1,6 @@
 import './addToRegistry'
 
+export * from './PortfolioMetadata'
 export * from './DetailedAccount'
 export * from './PortfolioCoin'
 export * from './Portfolio'
